+++ dtool_lookup_server_search_plugin_mongo/__init__.py
@@ -1,3 +1,0 @@
-"""dtool-lookup-server-search-plugin-mongo package."""
-
-__version__ = "0.2.0"