+++ CHANGELOG.rst
@@ -14,6 +14,9 @@ Added
 Changed
 ^^^^^^^
 
+- Replace ``setup.py`` by ``pyproject.toml``
+- Rebranded from dtool-lookup-server to dserver
+- Added sorting and pagination
 
 Deprecated
 ^^^^^^^^^^